apps/validator_engine/requirements.txt
# --- Core API ---
fastapi
uvicorn
python-multipart
pydantic

# --- AWS S3 + Rekognition ---
boto3
botocore

# --- Image Processing ---
Pillow
opencv-python

# --- EXIF Metadata ---
exifread

# --- Hashing & Duplicate Detection ---
ImageHash

# --- GPS Distance ---
haversine

# --- OCR for invoices ---
easyocr
torch
torchvision

# --- Redis for pHash store ---
redis
# --- Utilities ---
python-dotenv
requests
